refactored SearchBar.tsx
