+++ package.json
@@ -21,7 +21,6 @@
     "@defichain/playground-api-client": "^0.13.5",
     "@defichain/whale-api-client": "^0.18.6",
     "@headlessui/react": "^1.4.3",
-    "@popperjs/core": "^2.11.2",
     "@reduxjs/toolkit": "^1.7.1",
     "@tailwindcss/line-clamp": "^0.3.1",
     "bignumber.js": "^9.0.2",
@@ -38,7 +37,7 @@
     "react-dom": "17.0.2",
     "react-icons": "^4.3.1",
     "react-number-format": "^4.9.1",
-    "react-popper": "^2.2.5",
+    "@floating-ui/react-dom": "^0.4.3",
     "react-redux": "^7.2.6",
     "react-schemaorg": "^2.0.0",
     "recharts": "^2.1.8",
