+++ src/layouts/components/Header.tsx
@@ -7,7 +7,7 @@ import { MdClose, MdMenu } from 'react-icons/md'
 import { Container } from '@components/commons/Container'
 import { HeaderCountBar } from './HeaderCountBar'
 import { HeaderNetworkMenu } from './HeaderNetworkMenu'
-import { SearchBar } from '@components/commons/SearchBar'
+import { SearchBar } from '@components/commons/searchbar/SearchBar'
 
 export function Header (): JSX.Element {
   const [menu, setMenu] = useState(false)
